+++ src/main/resources/release_notes.md
@@ -26,6 +26,7 @@ Please do not forget to back up data.
 ### Rating
 
 ### Matches
+* reload of match details supports selection ranges (#1595)
 
 ### Lineup
 * SubstitutionEditor shows effect on loddarstats (#1626)